Added test for mutation
