+++ src/main.rs
@@ -1,4 +1,4 @@
-mod mutations;
+mod lib;
 
 fn main() {
     println!("Hello world!");
